fix(nav): drop native bottom tabs for JS tabs to kill the iOS tab-switch zoom
iOS 18+ plays a native UITabBarController cross-dissolve/zoom on every tab
switch (kept by iOS 26 Liquid Glass): the outgoing root-tab screen scaled into
the top-left corner and faded. Because react-native-bottom-tabs' SwiftUI
TabView is backed by that controller, the animation can't be disabled through
any public API — disablePageAnimations, a SwiftUI .transaction, and a
zero-duration UITabBarControllerDelegate all proved ineffective on-device
(upstream callstack/react-native-bottom-tabs#430).

Switch all platforms to the JS @react-navigation/bottom-tabs with the existing
custom BottomTabBar (what web already used, and what upstream Expensify does):

- BottomTabNavigator is now a single cross-platform component with
  animation: 'none', so tab switches are an instant cut everywhere. Deleted the
  redundant .web.tsx.
- BottomTabBar pads itself by the bottom safe-area inset so its items clear the
  iOS home indicator (no-op on web / Android without gesture nav).
- useBottomTabBarHeight collapses to a single index.ts returning 0 (the JS bar
  is a normal-flow sibling below the scene on every platform, not an overlay),
  removing the last react-native-bottom-tabs import. Deleted the iOS/Android
  variants.
- Removed the (ineffective) react-native-bottom-tabs patch and its details.md
  entry.

react-native-bottom-tabs / @bottom-tabs/react-navigation are now unused; their
removal from package.json + Podfile is a follow-up (needs a local pod install).
