+++ src/hooks/useBottomTabBarHeight/index.ts
@@ -1,14 +1,15 @@
 /**
- * Web: the custom JS tab bar (see `BottomTabBar`) is a normal-flow sibling
- * below the scene in `@react-navigation/bottom-tabs`' `BottomTabView` column,
- * NOT an overlay. The scene already ends at the top of the bar, so content
- * needs no extra clearance. Returning the bar height here would double-inset
- * every consumer (FAB/search offset, scroll `paddingBottom`, offline-indicator
- * margin), stacking a bar-sized dead band above the bar; this is the same bug
- * fixed on Android (see `index.android.ts`).
+ * The bottom tab bar (custom JS `BottomTabBar`) is a normal-flow sibling laid
+ * out *below* the scene by `@react-navigation/bottom-tabs`' `BottomTabView`, on
+ * every platform — NOT an overlay. The scene already ends at the top of the bar,
+ * so tab-root content needs no extra bottom clearance. Returning the bar height
+ * here would double-inset every consumer (FAB/search offset, scroll
+ * `paddingBottom`, offline-indicator margin), stacking a bar-sized dead band
+ * above the bar.
  *
- * iOS keeps the measured height (`index.native.ts`) because there the bar
- * genuinely overlays the scene.
+ * (Before moving off the native `react-native-bottom-tabs` bar, iOS returned a
+ * measured height here because that bar overlaid the scene. The JS bar does not
+ * overlay, so all platforms now return 0.)
  */
 function useBottomTabBarHeight(): number {
   return 0;
